FPV-43 migrated JWT token logic to page.py
